backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
sqlalchemy
pydantic
httpx
requests
gunicorn